pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Kevin Howbrook <kevin.howbrook@torchbox.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^6.2.4"
coverage = {extras = ["toml"], version = "^5.5"}
pytest-cov = "^2.12.1"
fabric = "^2.6.0"
black = "^21.7b0"
Pillow = "^8.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"